Extract sync diagnostics helpers
